src/requirements.txt
black==21.12b0
click==8.0.3
cycler==0.11.0
flake8==4.0.1
fonttools==4.28.5
kiwisolver==1.3.2
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
PyQt6==6.2.2
PyQt6-Qt6==6.2.2
PyQt6-sip==13.2.0
python-dateutil==2.8.2
PythonQwt==0.10.1
QtPy==2.0.0
scipy==1.7.3
six==1.16.0
tomli==1.2.3
typing_extensions==4.0.1
